firestore get and set working - signup and profile
