+++ src/app/profile/page.jsx
@@ -1,21 +1,24 @@
 "use client";
-import { useState, useEffect, useRef } from "react";
+import { useState, useEffect, useRef, use } from "react";
 import { Save, Upload, Camera } from "lucide-react";
-import { auth } from "@/config/firebase"; // Adjust the path to your firebase.js file
-import githubConfig from "@/config/githubConfig"; // Update import name
-import Header from "@/components/Header"; // Adjust the import path as necessary
+import { doc, getDoc, updateDoc } from "firebase/firestore";
+import { db, auth } from "@/config/firebase";
+import githubConfig from "@/config/githubConfig";
+import Header from "@/components/Header";
 import Image from "next/image";
+import { useRouter } from "next/navigation";
 
 
 const ProfilePage = () => {
+  const router = useRouter();
   const fileInputRef = useRef(null);
 
   const [formData, setFormData] = useState({
     name: "",
     email: "",
-    phone: "(555) 123-4567",
-    address: "123 Main Street, Springfield, USA",
-    company: "Doe Enterprises",
+    phone: "",
+    address: "",
+    company: "",
     position: "Manager",
   });
   const [userId, setUserId] = useState(""); // State to store the User ID
@@ -23,6 +26,8 @@ const ProfilePage = () => {
   const [config, setConfig] = useState(null); // State to store the GitHub config
   const [photoPreview, setPhotoPreview] = useState(null);
   const [isUploading, setIsUploading] = useState(false);
+
+
   // Fetch the User ID and generate the GitHub path
   useEffect(() => {
     const currentUser = auth.currentUser;
@@ -40,7 +45,6 @@ const ProfilePage = () => {
         name: username, // Set name from Firebase username
         email: email // Set email from Firebase
       }));
-      fetchProfileData(username);
 
     }
   }, []);
@@ -53,24 +57,43 @@ const ProfilePage = () => {
     });
   };
 
-  const fetchProfileData = async (username) => {
-    try {
-      const response = await fetch(`/api/profile/${username}`);
-      if (response.ok) {
-        const data = await response.json();
-        setFormData(prev => ({ ...prev, ...data }));
-        if (data.profilePhoto) {
-          setPhotoPreview(data.profilePhoto);
+
+  // Fetch profile data from Firestore when the component mounts
+  useEffect(() => {
+    const fetchProfileFromFirestore = async () => {
+      const currentUser = auth.currentUser;
+      if (currentUser) {
+        const uid = currentUser.uid;
+        const email = currentUser.email;
+        const username = email.split("@")[0];
+        setUserId(uid);
+        setGithubPath(`${username}-${uid}`);
+        const docId = email.replace(/\./g, "_");
+        const userDoc = await getDoc(doc(db, "users", docId));
+        if (userDoc.exists()) {
+          const data = userDoc.data();
+          const userData = data.user || {};
+          setFormData({
+            name: userData.displayName || "",
+            email: userData.email || email,
+            phone: userData.phone || "",
+            address: userData.address || "",
+            company: userData.company || "",
+            position: userData.position || "Manager",
+          });
+          if (userData.photoURL) setPhotoPreview(userData.photoURL);
+        } else {
+          setFormData(prev => ({ ...prev, email }));
         }
       }
-    } catch (error) {
-      console.error("Error loading profile data:", error);
-    }
-  };
+    };
+    fetchProfileFromFirestore();
+  }, []);
 
   const handlePhotoClick = () => {
     fileInputRef.current?.click();
   };
+
   const handlePhotoChange = async (e) => {
     const file = e.target.files?.[0];
     if (file) {
@@ -108,26 +131,31 @@ const ProfilePage = () => {
   const handleSave = async (e) => {
     e.preventDefault();
     try {
-      const response = await fetch(`/api/profile/save`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          data: formData,
-          path: githubPath
-        })
-      });
-
-      if (response.ok) {
-        alert("Profile updated successfully!");
+      const currentUser = auth.currentUser;
+      if (!currentUser) {
+        alert("User not authenticated");
+        return;
       }
+      const docId = currentUser.email.replace(/\./g, "_");
+      await updateDoc(doc(db, "users", docId), {
+        user: {
+          uid: userId,
+          email: currentUser.email,
+          displayName: formData.name,
+          photoURL: photoPreview || currentUser.photoURL,
+          phone: formData.phone,
+          address: formData.address,
+          company: formData.company,
+          position: formData.position,
+        }
+      });
+      alert("Profile updated successfully!");
+      router.push("/dashboard");
     } catch (error) {
       console.error("Error saving profile:", error);
       alert("Failed to update profile");
     }
   };
-
